app/requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.8.0
ecdsa==0.19.0
fastapi==0.115.6
idna==3.10
jose==1.0.0
pyasn1==0.6.1
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
python-jose==3.3.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing==3.7.4.3
typing_extensions==4.12.2
